String+ContentValidation: Work around a regression in the NSPredicate parser (FB9802649)
Starting with iOS15, the predicate `SELF MATCHES <RegEx>` crashes for the (valid) regular expression
`^((\+)|(00))[0-9]{6,14}|[0-9]{6,14}$`. This has worked in older operating systems.
This is now refactored to not use `NSPredicate` anymore. The nice byproduct is that it now works
on Linux as well, since `NSPredicate` is not (yet?) implemented in swift-corelibs-foundation.