Logging: Add LogFile, a simple (textual) log file abstraction.
Logging happens on background queues, so it should be save to call
concurrently (and possibly even from performance critical code).

NOTE: If maximum performance is a requirement, we could rework this to coalesce
the input to not hit the filesystem less often. Then again,
this would probably less crash-safety, since we would buffer more.
As always, it's a tradeoff. That said, it might be interesting to
enhance this class with a `strategy` parameter that can switch
between both behaviors.