RadixInterpolation: Split the collection formatter into two different methods.
WARNING: This is a breaking API change, I'm sorry for that, but it improves
clarity as previously it could have lead to a "wrong" output when you used the
default arguments. This way it's more clear what happens and whether you want
to format the collection as a bunch of single values or a huge combined value.